FwdFPGA: implement basic XDMA transfer functionality
This emulates the dedicated FPGA DRAM using a temporary file, and implements
reading and writing to/from the DRAM via the XDMA functionality. I tested the
FwdFPGA Driver initialization with it (which writes DTA programs to the DRAM),
as well as the XDMA Driver test partition (which writes and reads "Hello world").